+++ minime/views.py
@@ -1,4 +1,4 @@
-from django.http import HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
 from rest_framework import generics
 from minime import forms, models, serializers, utils
@@ -14,6 +14,10 @@ def index(request):
     return render(request, "index.html", {"form": forms.ShortenURLForm})
 
 
+def ping(request):
+    return HttpResponse(status=200)
+
+
 def redirect(request, hash):
     """
     View that redirects to `Url.url` if it exists.
@@ -26,10 +30,8 @@ def redirect(request, hash):
     url = utils.cache_get_short_url(hash)
 
     if not url:
-        urlobj = get_object_or_404(
-            models.Url, hash=hash
-        )  # gets object, if not found returns 404 error
-        url = urlobj.url
+        urlobj = get_object_or_404(models.Url, hash=hash)
+        url = getattr(urlobj, "url")
         utils.cache_set_url(hash, url)
 
     utils.create_stats(request, hash)